\name{StockIndex}
\alias{StockIndex}
\docType{data}
\title{
  Stock Index Data
}
\description{
  Month-end price data of six stock indices.
}

\usage{data(StockIndex)}

\format{
  A data frame with 240 month-end observations of six stock indices: SP
  500, Nikkei 225, FTSE 100, CAC40, GDAX and Hang Seng index. The sample
  starts at 1991-07-31 and ends in  2011-06-30.      
}

\details{
  The data set has been obtained from Yahoo Finance and hereby the
  unadjusted closing prices have been retrieved. 
}

\source{
  \url{https://finance.yahoo.com/}
}

\examples{
data(StockIndex)
}
\keyword{datasets}
