\name{StockIndexAdj}
\alias{StockIndexAdj}
\docType{data}
\title{
  Stock Index Data
}
\description{
  Adjusted month-end price data of six stock indices.
}

\usage{data(StockIndexAdj)}

\format{
  A data frame with 240 month-end observations of six stock indices: SP
  500, Nikkei 225, FTSE 100, CAC40, GDAX and Hang Seng index. The sample
  starts at 1991-07-31 and ends in  2011-06-30.      
}

\details{
  The data set has been obtained from Yahoo Finance and hereby the
  adjusted closing prices have been retrieved. 
}

\source{
  \url{https://finance.yahoo.com/}
}

\examples{
data(StockIndexAdj)
}
\keyword{datasets}
